requirements.txt
requests
rich
python-dateutil
python-dotenv
pytz
websockets